ml-models/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
Pillow==10.3.0
numpy==1.26.4
opencv-python-headless==4.10.0.82
ultralytics==8.2.28
onnxruntime==1.18.0
lightgbm==4.3.0
scikit-learn==1.5.0
pandas==2.2.2
httpx==0.27.0
python-dotenv==1.0.1
# Background removal — strips the scene background so the damage model
# only fires on the actual device surface, not the floor / table / walls.
rembg==2.0.57
